add test

+++ src/main/java/co/simplon/dnd_heroic_battle_api/dtos/monsters/MonsterHpUpdateDto.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record MonsterHpUpdateDto(@Positive Long monsterId, @Positive int amount, @NotNull String type,
-                                 Integer damageTypeId) {
+                                 Long damageTypeId) {
 }
